+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/coveragedata/FastArrayCodeProvider.java
@@ -76,13 +76,13 @@ public class FastArrayCodeProvider extends AbstractCodeProvider
 
 	@SuppressWarnings("deprecation")
 	public void generateCodeThatIncrementsCoberturaCounter(
-			MethodVisitor nextMethodVisitor, Integer counterId, String className) {
+			MethodVisitor nextMethodVisitor, int counterId, String className) {
 		/*cobertura_counters[value('lastJumpIdVariableIndex')]++;*/
 		/*cobertura_counters.*/
 		nextMethodVisitor.visitFieldInsn(Opcodes.GETSTATIC, className,
 				COBERTURA_COUNTERS_FIELD_NAME, COBERTURA_COUNTERS_FIELD_TYPE);
 		/*index:*/
-		nextMethodVisitor.visitLdcInsn((int) counterId);
+		nextMethodVisitor.visitLdcInsn(counterId);
 		nextMethodVisitor.visitInsn(Opcodes.DUP2);
 		nextMethodVisitor.visitInsn(Opcodes.IALOAD);
 		nextMethodVisitor.visitLdcInsn(1);
@@ -92,7 +92,7 @@ public class FastArrayCodeProvider extends AbstractCodeProvider
 		if (collectExecutionTrace) {
 			// add the statement to the execution trace... TODO
 			nextMethodVisitor.visitLdcInsn(className);
-			nextMethodVisitor.visitLdcInsn((int) counterId);
+			nextMethodVisitor.visitLdcInsn(counterId);
 			nextMethodVisitor.visitMethodInsn(Opcodes.INVOKESTATIC, Type
 					.getInternalName(ExecutionTraceCollector.class), "addStatementToExecutionTrace",
 					"(Ljava/lang/String;I)V");
